arreglos de tipado

+++ src/app/api/company/[id]/route.ts
@@ -2,8 +2,14 @@ import { NextResponse } from "next/server";
 import prisma from "@/lib/db";
 export async function GET(
   req: Request,
-  { params }: { params: { id: string } },
+  context: { params: Promise<{ id: string }> }, // params es Promise
 ) {
+  const params = await context.params;
+  const id = Number(params.id);
+  if (isNaN(id)) {
+    return NextResponse.json({ error: "ID inválido" }, { status: 400 });
+  }
+
   const company = await prisma.company.findUnique({
     where: { id: Number(params.id) },
   });
@@ -13,10 +19,15 @@ export async function GET(
 
 export async function PUT(
   req: Request,
-  { params }: { params: { id: string } },
+  context: { params: Promise<{ id: string }> }, // params es Promise
 ) {
   try {
+    const params = await context.params;
     const id = Number(params.id);
+    if (isNaN(id)) {
+      return NextResponse.json({ error: "ID inválido" }, { status: 400 });
+    }
+
     const body = await req.json();
 
     const updated = await prisma.company.update({
@@ -38,8 +49,14 @@ export async function PUT(
 
 export async function DELETE(
   req: Request,
-  { params }: { params: { id: string } },
+  context: { params: Promise<{ id: string }> }, // params es Promise
 ) {
+  const params = await context.params;
+  const id = Number(params.id);
+  if (isNaN(id)) {
+    return NextResponse.json({ error: "ID inválido" }, { status: 400 });
+  }
+
   try {
     await prisma.company.delete({
       where: { id: Number(params.id) },
